merge final feign progress

+++ IInd Layer Services/Microservice_SmartFood_FeignProgress/src/main/java/com/base_package/model/CustomerTrackResponseModel.java
@@ -11,5 +11,5 @@ public class CustomerTrackResponseModel {
 	private String utUuid;
 	private String tackingDate;
 	private String mealTime;
-	private int calories;
+	private Float calories;
 }
